+++ scripts/send-bugsnag-symbols.js
@@ -1,29 +1,34 @@
-import chalk from 'chalk';
-import {exec} from 'child_process';
-import {positive, error} from './utils';
+import chalk from 'chalk'
+import {exec} from 'child_process'
+import {positive, error} from './utils'
 
-const packageInfo = require('../package.json');
+const version = require('../package.json').version
+// buildID is the first command line argument, if supplied
+const buildID = process.argv.length >= 3 ? process.argv[2] : null
+const appVersion = buildID ? `${version} (${buildID})` : version
 
-const {version} = packageInfo;
+console.log(`appVersion: ${appVersion}`)
+console.log(chalk.green('Bundling react-native assets...'))
 
-console.log(chalk.green('Bundling react-native assets...'));
+exec(
+  'react-native bundle --platform ios --entry-file index.ios.js --dev false --bundle-output ./ios/main.jsbundle --sourcemap-output ./sourcemap.js',
+  (err, stdout) => {
+    console.log('RN bundle result:', positive(stdout), error(err))
 
-exec('react-native bundle --platform ios --entry-file index.ios.js --dev false --bundle-output ./ios/main.jsbundle --sourcemap-output ./sourcemap.js', (err, stdout) => {
-  console.log('RN bundle result:', positive(stdout), error(err));
+    if (err) return
 
-  if (err) return;
-
-  // upload symbol files to bugsnag
-  exec(
-    `curl https://upload.bugsnag.com/ \
+    // upload symbol files to bugsnag
+    exec(
+      `curl https://upload.bugsnag.com/ \
         -F apiKey=f108fb997359e5519815d5fc58c79ad3 \
-        -F appVersion=${version} \
-        -F minifiedUrl="main.jsbundle" \
+        -F "appVersion=${appVersion}" \
+        -F minifiedUrl=main.jsbundle \
         -F sourceMap=@./sourcemap.js \
         -F minifiedFile=@./ios/main.jsbundle \
         -F overwrite=true`,
-    (err1, stdout1) => {
-      console.log('Bugsnag deploy result:', positive(stdout1), error(err1));
-    },
-  );
-});
+      (err1, stdout1) => {
+        console.log('Bugsnag deploy result:', positive(stdout1), error(err1))
+      }
+    )
+  }
+)