+++ src/components/Notifications.tsx
@@ -8,7 +8,9 @@ import EventCard from './event-cards/EventCard'
 import ListFooter from './ListFooter'
 import {navBarStyle, placeholderStyle} from './styles'
 import {colors} from 'src/constants'
-import {View} from 'react-native'
+import {View, StyleSheet} from 'react-native'
+import SwitchButton from './SwitchButton'
+import {PINK, WHITE} from '../constants/colors'
 
 type Props = {
   wocky?: IWocky
@@ -17,12 +19,15 @@ type Props = {
 @inject('wocky')
 @observer
 class Notifications extends React.Component<Props> {
+  componentWillMount() {
+    this.props.wocky!.notifications.setMode(1)
+  }
   componentDidMount() {
-    this.props.wocky!.viewNotifications()
+    this.props.wocky!.notifications.readAll()
   }
 
   componentWillUnmount() {
-    this.props.wocky!.viewNotifications()
+    this.props.wocky!.notifications.readAll()
   }
 
   render() {
@@ -34,34 +39,54 @@ class Notifications extends React.Component<Props> {
       <DraggablePopupList
         fadeNavConfig={{
           back: true,
-          title: <RText style={navBarStyle.titleStyle}>Updates</RText>,
+          title: <RText style={navBarStyle.titleStyle}>{notifications.title}</RText>,
         }}
         headerInner={
-          <RText size={17} weight={'Medium'}>
-            Updates
-          </RText>
+          <View style={{flex: 1, alignItems: 'center'}}>
+            <SwitchButton
+              text1="Updates"
+              text2="Requests"
+              switchWidth={220}
+              switchHeight={32}
+              btnStyle={{}}
+              btnHeight={32}
+              btnBorderColor={PINK}
+              btnBackgroundColor={PINK}
+              switchBackgroundColor={WHITE}
+              switchBorderColor={PINK}
+              switchBorderRadius={16}
+              activeFontColor={WHITE}
+              fontColor={PINK}
+              onValueChange={index => {
+                notifications.setMode(index)
+                notifications.load()
+              }}
+            >
+              {notifications.hasUnreadRequests && <View style={styles.newDot} />}
+            </SwitchButton>
+          </View>
         }
-        data={notifications.length > 0 ? notifications.list.slice() : []}
+        data={notifications.data}
         renderItem={this.renderItem}
         keyExtractor={this.keyExtractor}
         onEndReachedThreshold={0.5}
         onEndReached={() => notifications.load()}
         ListFooterComponent={observer(() => (
           <View>
-            {notifications.length === 0 ? (
+            {notifications.data.length === 0 ? (
               <View>
                 <RText
                   weight="Regular"
                   color={colors.GREY}
                   style={placeholderStyle.placeholderText as any}
                 >
-                  No new updates
+                  {notifications.emptyTitle}
                 </RText>
               </View>
             ) : (
               <ListFooter
                 style={{backgroundColor: 'white'}}
-                finished={notifications.finished || notifications.length === 0}
+                finished={notifications.finished || notifications.data.length === 0}
               />
             )}
           </View>
@@ -74,5 +99,19 @@ class Notifications extends React.Component<Props> {
 
   keyExtractor = (item: any) => item.id
 }
+
+const styles = StyleSheet.create({
+  newDot: {
+    position: 'absolute',
+    top: -5,
+    right: 0,
+    borderColor: 'white',
+    borderWidth: 2,
+    borderRadius: 13 / 2,
+    width: 13,
+    height: 13,
+    backgroundColor: colors.GOLD,
+  },
+})
 
 export default Notifications