+++ third-party/wocky-client/src/model/Event.ts
@@ -1,17 +1,31 @@
-import {types, Instance} from 'mobx-state-tree'
+import {types, Instance, getParent} from 'mobx-state-tree'
 import {Timeable, ITimeableData} from './Timeable'
 import {Base, IBaseData} from './Base'
 import {IProfile} from './Profile'
+import {flow} from 'mobx'
 
 export const Event = types
-  .compose(Base, Timeable, types.model({id: types.identifier}))
+  .compose(Base, Timeable, types.model({id: types.identifier, unread: true}))
   .named('Event')
   .views(() => ({
     get target(): IProfile {
       throw new Error('Abstract method!')
     },
   }))
-  .actions(() => ({
+  .actions(self => ({
+    read: () => {
+      self.unread = false
+    },
+    remove: flow(function*() {
+      const parent: any = getParent(getParent(self))
+      yield self.transport.notificationDelete(self.id)
+      parent.remove(self.id)
+    }),
+    removeAfterDelay: flow(function*(seconds: number) {
+      const parent: any = getParent(getParent(self))
+      yield self.transport.notificationDelete(self.id)
+      setTimeout(() => parent.remove(self.id), seconds * 1000)
+    }),
     process: () => {
       // do nothing
     },